requirements.txt
streamlit
streamlit-chat
streamlit_extras
python-dotenv
langchain
openai
langchain-community